backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart
python-dotenv
